refactor(foldrun): move A2A Cloud Run service to src/foldrun-a2a
Move A2A service files from foldrun-agent/ to src/foldrun-a2a/
alongside other Cloud Run services. The deploy script now syncs
foldrun_app/ from the agent directory at build time.

Moved files:
- main.py, Procfile, .gcloudignore -> src/foldrun-a2a/
- a2a_agent_card.py -> src/foldrun-a2a/ (top-level import)
- deploy_a2a_cloudrun.sh -> src/foldrun-a2a/deploy.sh
- register_a2a_gemini.sh -> src/foldrun-a2a/register_gemini.sh
- test_a2a.py -> src/foldrun-a2a/
