+++ applications/foldrun/src/foldrun-a2a/.gcloudignore
@@ -1,0 +1,11 @@
+.gcloudignore
+.git
+.gitignore
+.venv
+__pycache__
+*.pyc
+uv.lock
+deploy.sh
+deploy_generic.sh
+register_gemini.sh
+test_a2a.py
