+++ applications/foldrun/src/foldrun-a2a/deploy_generic.sh
