+++ applications/foldrun/src/foldrun-a2a/register_gemini.sh
